added a drop rule for ssh rate limiting
For the ssh connections over tor after the first 3 packets with the a NEW flag are accepted by iptables the rate limiting of 3/min starts. Any packets that are above that threshold will not match the rate limiting rule and will be processed by the next rule in the list.

The next rule is a drop rule specifically for tcp packets with the NEW flag from the tor user to the local loopback.

All the traffic from the tor user to destination port 22 on the local loobpack above the rate limit threshold will match this drop rule.

Here is the output of the counters showing of 7 ssh connections attempted over tor.

```
[6:360] -A OUTPUT -o lo -p tcp -m tcp --dport 22 -m owner --uid-owner 109 -m state --state NEW -m limit --limit 3/min --limit-burst 3 -m comment --comment "Rate limit traffic from tor to the ssh server" -j ACCEPT
[28:1680] -A OUTPUT -o lo -p tcp -m tcp --dport 22 -m owner --uid-owner 109 -m state --state NEW -m comment --comment "Drop all other new conn$ctions from tor to ssh server" -j LOGNDROP
[1138:91786] -A OUTPUT -o lo -p tcp -m tcp --dport 22 -m owner --uid-owner 109 -m state --state RELATED,ESTABLISHED -m comment --comment "Allo$ the established traffic from tor to ssh server" -j ACCEPT
[777:674290] -A OUTPUT -p tcp -m owner --uid-owner 109 -m state --state NEW,RELATED,ESTABLISHED -m comment --comment "tor instance that provid$s ssh access" -j ACCEPT
[0:0] -A OUTPUT -m owner --uid-owner 109 -m comment --comment "Drop all other traffic for the tor instance used for ssh" -j LOGNDROP
14 minutes ago - Edit - Delete
```

You can see the counters increase for the first couple ssh connections on the first line which is the ssh rate limitng rule.

The second line is the ssh drop rule for new packets. The counters increased for that after the rate limiting threshold was reached.

This does not affect the staging environments unless testing over tor.
This does not affect the development at all.