iptable interface restrictions
Since we are now using the anisble built in fact to select the correct interface name, we can add restrictions for the interface to the majority of the iptable rules.

This allows restricting what applicaitons communicate to the apache vhost and redis from the localhost.

The app and monitor templated rules were updated and so were the testing allow direct access exemptions to look like the production rules.

This will cause a merge conflict when merging the `0.3.4` release back into develop.