Reduce OSSEC spam for deb pkg updates
For security updates installed by cron-apt OSSEC has been alerting on:

* 1 grouped email for packages installed as part of cron-apt
* Indivual emails (up to the throttle limit) for each file changed
* 1 Daily report summerizing any files changed that day.

This pr adds a custom group `sec_update_spam` for the related types of events that occur during the cron-apt upgrade window 04:00-05:15.

A window 15 min past when cron-apt can run was chosen to provide time for packages to be downloaded and installed if the time selected by cron-apt was at the end of the window.

The custom group sec_update_spam related alerts will all be grouped together and sent as 1 alert.

The daily report will still include syscheck event outside and during the the cron-apt window and will also include the related dpkg events that occured during the update window.

Packages that are installed outside the cron-apt window will not be included in the syscheck daily report.